+++ vservice/test/thrift/thrift_test.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"git.apache.org/thrift.git/lib/go/thrift"
 	"testing"
-	"time"
-	"vtool/idl/thrift/gen-go/thriftError"
 	clientCommon "vtool/vservice/client/common"
 	clientThrift "vtool/vservice/client/rpc_client"
 	"vtool/vservice/common"
@@ -19,7 +17,7 @@ type helloServiceHandler struct {
 
 func (h *helloServiceHandler) SayHello(req *SayHelloReq) (*SayHelloRes, error) {
 	return &SayHelloRes{
-		Data: &SayHelloData{Val: "this is val"},
+		Data: &SayHelloData{Val: "this is thrift val"},
 	}, nil
 }
 
@@ -39,41 +37,6 @@ func TestThriftServer(t *testing.T) {
 		return
 	}
 
-}
-
-var thriftClient common.RpcClient
-
-func rpc(ctx context.Context, hashKey string, timeout time.Duration, fn func(*TestServiceClient) error) error {
-	return thriftClient.Rpc(ctx, &common.ClientCallerArgs{
-		Lane:    "",
-		HashKey: hashKey,
-		TimeOut: timeout,
-	}, func(c interface{}) error {
-		ct, ok := c.(*TestServiceClient)
-		if ok {
-			return fn(ct)
-		} else {
-			return fmt.Errorf("reflect client rpc_client error")
-		}
-	})
-}
-
-func SayHello(ctx context.Context, req *SayHelloReq) (res *SayHelloRes) {
-	err := rpc(ctx, "", time.Millisecond*3000,
-		func(c *TestServiceClient) (e error) {
-			res, e = c.SayHello(req)
-			return e
-		})
-
-	if err != nil {
-		res = &SayHelloRes{
-			ErrInfo: &thriftError.ErrInfo{
-				Code: -1,
-				Msg:  fmt.Sprintf("rpc service:%s serv:%s method:SayHello err:%v", "censor", common.Thrift, err),
-			},
-		}
-	}
-	return
 }
 
 func TestNewThriftClient(t *testing.T) {
